Re-enable bulletproof transactions
Disabling bulletproofs caused problems with payment to multiple
recipients (it appears to be caused by wallets trying to spend funds
that originally was part of a bulletproof).

The 0.1.2.4 released forced bulletproofs off to try to remain
compatible with 0.1.2.3, but this didn't work: there was already a
bulletproof transaction on the network that caused 0.1.2.4 daemons to
fork from 0.1.2.3 daemons.

Thus the current network essentially requires a 0.1.2.4 daemon to sync
properly, and since 0.1.2.4 already support bulletproofs, there is no
harm in turning them back on (they were, after all, running on the
network before the 0.1.2.2 release all the way back to block 100).