Restore licence to BSD
While I'm a big fan of the GPL, changing the licence on a monero fork is
problematic: while it is possible for BSD code to go into a GPL project,
the other way (i.e. returning a GPL-licensed fix to BSD code) isn't
compatible.

This commit returns the license to Monero's BSD license.