Add files via upload
Updated description removing the term "electrum seed" from user prompt on restore deterministic wallet using the seed to changed as "25 word mnemonic seed" as the primary mnemonic seed key that determines the wallet. Not to be confused with the additional passphrase 25 word mnemonic seed for encrypting the first mnemonic seed .  No changes were made to the vars or global or cvars, just the text description to the user. Removes confusion to those using similar sounding coins like "electroneum" or "electronero" or "monero:".