p2p: do not send last_seen timestamp to peers
This can be used for fingerprinting and working out the network topology.

Instead of sending the first N (which are sorted by last seen time), we sent a random subset of the first N+N/5, which ensures reasonably recent peers are used, while preventing repeated calls from deducing new entries are peers the target node just connected to.

The list is also randomly shuffled so the original set of timestamps cannot be approximated.

Co-Authored-By: moneromooo-monero <moneromooo-monero@users.noreply.github.com>
Co-Authored-By: Tom Smeding <tomsmeding@users.noreply.github.com>